# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Movie Agent API",
    "description": "Unified REST and MCP server for local UK trending movie database",
    "version": "1.0.0"
  },
  "paths": {
    "/search_trending_movies": {
      "get": {
        "summary": "Search Trending Movies",
        "operationId": "search_trending_movies_search_trending_movies_get",
        "parameters": [
          {
            "name": "genre",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by genre ('action', 'comedy', 'drama', 'horror', 'scifi', 'thriller', etc.)",
              "title": "Genre"
            },
            "description": "Filter by genre ('action', 'comedy', 'drama', 'horror', 'scifi', 'thriller', etc.)"
          },
          {
            "name": "min_rating",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "description": "Minimum IMDb score (default 6.0)",
              "default": 6.0,
              "title": "Min Rating"
            },
            "description": "Minimum IMDb score (default 6.0)"
          },
          {
            "name": "platform",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Streaming platform ('Netflix', 'Amazon Prime Video', 'Disney Plus', 'Apple TV Plus', etc.)",
              "title": "Platform"
            },
            "description": "Streaming platform ('Netflix', 'Amazon Prime Video', 'Disney Plus', 'Apple TV Plus', etc.)"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "description": "Maximum records to return",
              "default": 20,
              "title": "Limit"
            },
            "description": "Maximum records to return"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/list_available_platforms": {
      "get": {
        "summary": "List Available Platforms",
        "operationId": "list_available_platforms_list_available_platforms_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/get_subscription_pricing": {
      "get": {
        "summary": "Get Subscription Pricing",
        "operationId": "get_subscription_pricing_get_subscription_pricing_get",
        "parameters": [
          {
            "name": "region",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "GB",
              "title": "Region"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/sse": {
      "get": {
        "summary": "Handle Sse",
        "operationId": "handle_sse_sse_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/messages/": {
      "post": {
        "summary": "Handle Messages",
        "operationId": "handle_messages_messages__post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}